chore: drop old implementation
