+++ src/index.js
@@ -1,7 +1,5 @@
 import ValidationProvider from './components/ValidationProvider.js';
 import ValidationField from './components/ValidationField.js';
-import ValidationProviderNew from './components/ValidationProviderNew.js';
-import ValidationFieldNew from './components/ValidationFieldNew.js';
 import ValidationFieldArray from './components/ValidationFieldArray.js';
 import ValidationErrors from './components/ValidationErrors.js';
 import { register } from './validators.js';
@@ -11,8 +9,6 @@ export {
   ValidationProvider,
   ValidationField,
   ValidationFieldArray,
-  ValidationProviderNew,
-  ValidationFieldNew,
   ValidationErrors,
   register as registerValidator,
   symbols
